+++ index.html
@@ -230,6 +230,6 @@
   </template>
 
   <!-- Скрипты -->
-  <script type="text/javascript" src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
